requirements.txt
asgiref==3.11.1
coverage==7.13.5
Django==6.0.3
djangorestframework==3.17.0
freezegun==1.5.5
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
